+++ db/migrate/20170308104405_create_works.rb
@@ -3,7 +3,7 @@ class CreateWorks < ActiveRecord::Migration[5.0]
     create_table :works do |t|
       t.references :user, foreign_key: true
       t.references :month, foreign_key: true
-      t.decimal :power
+      t.integer :power
 
       t.timestamps
     end
